requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
httpx==0.28.1
trafilatura==2.0.0
redis[hiredis]==5.2.1
pyyaml==6.0.2
pydantic==2.10.4
pydantic-settings==2.7.1
duckduckgo-search==7.5.3
gunicorn==23.0.0
slowapi==0.1.9
structlog==24.4.0
sse-starlette==2.2.1
flashrank==0.2.9
readability-lxml==0.8.1
circuitbreaker==2.0.0

# Dev / Testing
pytest==8.3.4
pytest-asyncio==0.25.0
